% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{boot_mse}
\alias{boot_mse}
\title{Perform a semi-parametric bootstrap}
\usage{
boot_mse(object, B = 2000, seed = 1)
}
\arguments{
\item{object}{an object created with the function \code{\link{mse}}.}

\item{B, }{the number of bootstrap replications. Defaults to 2000.}

\item{seed}{random seed for reproducibility. Defaults to 1.}
}
\value{
A matrix with the observed (and latent) variables
and the fitted frequencies of the \code{B} bootstrap samples in the columns.
}
\description{
\code{boot_mse} Draws random multinomial bootstrap samples of
size N from the observed frequencies, including those of cells with missing values,
and the fitted frequencies of the cells with structural zeros, and fits the model
to the bootstrap samples excluding the cells with structural zeros.
}
\examples{
# Bootstrap the object:
AB <- mse(Freq ~ A*b + B*a + a*b, dframe = NewZealand[, c(1:2, 5:6, 9)])
\dontrun{
boot_AB <- boot_mse(AB)
}
}
