% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NewZealand}
\alias{NewZealand}
\title{New Zealand population registers.}
\format{
Data frame with 1296 cases and 9 variables:
\describe{
  \item{A}{census (0 = not observed, 1  = observed)}
  \item{B}{dia (0 = not observed, 1  = observed)}
  \item{C}{moh (0 = not observed, 1  = observed)}
  \item{B}{moe (0 = not observed, 1  = observed)}
  \item{a}{self-reported ethnicity in census (0 = non-Moari, 1  = Maori, NA = not reported)}
  \item{b}{self-reported ethnicity in dia (0 = non-Moari, 1  = Maori, NA = not reported)}
  \item{c}{self-reported ethnicity in moh (0 = non-Moari, 1  = Maori, NA = not reported)}
  \item{d}{self-reported ethnicity in moe (0 = non-Moari, 1  = Maori, NA = not reported)}
  \item{Freq}{The observed frequency}
}
}
\source{
Stats New Zealand.
}
\usage{
NewZealand
}
\description{
A contingency table containing self-reported classifications
of ethnicity (Maori vs non-Maori) of in four incomplete population
registers.
}
\details{
Cells for which A, B, C, and D is 0 are not observed in any of the four registers,
and constitute structural zeros.
}
\keyword{datasets}
